# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.SignUpResp:
    properties:
      username:
        type: string
    type: object
  model.Category:
    properties:
      icon:
        example: https://example.com/example.png
        type: string
      id:
        type: integer
      name:
        example: Soccer
        type: string
      type:
        example: sport
        type: string
    type: object
  util.HTTPError:
    properties:
      error:
        example: status bad request
        type: string
    type: object
info:
  contact:
    email: timadinorth@gmail.com
    name: Tim Adi
  description: Betting exchange API documentation
  title: BetPub exchange API
  version: "1.0"
paths:
  /categories:
    get:
      description: Returns list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Category'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.HTTPError'
      summary: Get categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Creates new category and assigns unique Id
      parameters:
      - description: Category name
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: Icon path
        in: body
        name: icon
        required: true
        schema:
          type: string
      - description: Category type
        in: body
        name: type
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.HTTPError'
      summary: Add a category
      tags:
      - categories
  /signin:
    post:
      consumes:
      - application/json
      description: Create new session for user
      parameters:
      - description: Username
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: Password plan text
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.HTTPError'
      summary: User signin
      tags:
      - auth
  /signout:
    post:
      description: Delete user session
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.HTTPError'
      summary: User signout
      tags:
      - auth
  /signup:
    post:
      consumes:
      - application/json
      description: Creates new user
      parameters:
      - description: Username
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: Password plan text
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.SignUpResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.HTTPError'
      summary: Register user
      tags:
      - auth
swagger: "2.0"